fix: address bugbot feedback — hybrid ERROR propagation, dep upgrade, coverage
- _run_hybrid: treat CheckStatus.ERROR as non-passing (deno_failed/
  node_failed now check FAILED|ERROR); propagate ERROR aggregate
  status when either sub-result errored rather than silently passing
- cryptography: bump floor to >=46.0.6 (GHSA-m959-cc7f-wv43)
- tests: add 7 hybrid/auto_fix/target-dirs tests for lint_format.py
  to lift diff coverage above 80% threshold
- refactor: move TestJavaScriptExpectCheck (464 lines) to new file
  test_javascript_eslint_expect.py to satisfy code-sprawl gate
